requirements.txt
pymem
customtkinter
Pillow
pyinstaller
winshell
pywin32
